Completed initial port of BIST.

+++ examples/bist_test/routines.h
@@ -16,7 +16,7 @@ char* hex_decode(const char *in, char len,char *out);
 unsigned short read_ascii_hex_u16(void);
 void adi_hw_reset(void);
 char *xil_fgets(char *buf, int bsize);
-float standard_deviation(float data[], int n);
+int standard_deviation(int data[], int n);
 int rsscanf(char* input_string, const char* format, ...);
 #define sscanf rsscanf
 double sqrt(double val);
